+++ src/propsheet/TerminalPropsheetPage.h
