+++ src/propsheet/propsheet.vcxproj
@@ -23,7 +23,7 @@
     <ClCompile Include="LayoutPage.cpp" />
     <ClCompile Include="ColorsPage.cpp" />
     <ClCompile Include="ColorControl.cpp" />
-    <ClCompile Include="TerminalPage.cpp" />
+    <ClCompile Include="TerminalPropsheetPage.cpp" />
     <ClCompile Include="registry.cpp" />
     <ClCompile Include="util.cpp" />
     <ClCompile Include="precomp.cpp">
@@ -41,7 +41,7 @@
     <ClInclude Include="LayoutPage.h" />
     <ClInclude Include="ColorsPage.h" />
     <ClInclude Include="ColorControl.h" />
-    <ClInclude Include="TerminalPage.h" />
+    <ClInclude Include="TerminalPropsheetPage.h" />
     <ClInclude Include="menu.h" />
     <ClInclude Include="precomp.h" />
   </ItemGroup>
